Release v1.3.0.
## Segment in one channel, analyse in all.

Segmentation is done only on one channel, specified by the user
e.g. on the GUI. The intensity measurements are done on all
channels using the circles found in the segmentation channel.

## Limit the number of detected circles.

A new settings in the GUI (in the 'Advanced parameters panel')
lets the user select a maximal number of detected circles. Only
the N first circles are then returned, by increasing sensitivity
value.